Move ecdf utility functions to stats module
